Refactor Share view into a Share/ module directory
Split the ~3.3k-line views/Share.tsx into views/Share/ — an orchestrator (index.tsx) plus the six step components (Queue/Redact/Review/Package/Submit/Done) and shared types/helpers/styles/UI modules. Behavior-preserving; also drops four dead props surfaced by review (QueueStep.sessionById, ReviewStep.onHelp, SubmitStep.onBack, and the vestigial DoneStep error-state chain).

Supporting changes bundled in: export ApiError from api.ts (consumed by SubmitStep); surface supported_institution_email_policy in the daemon share-destination payload; add AGENTS.md (CLAUDE.md mirror) and simplified_upload.md design notes.
